Introduced ProcessLauncher to the project.
Also moved all the launcher classes from tornadoweb to the launcher
module and fixed test reference.

Refs: #249
