+++ requirements/basic.txt
@@ -1,3 +1,4 @@
 pyyaml==3.13
 tornado==5.1.1
 six==1.11.0
+pexpect==4.6.0
